Implementação do CDU Principal: Cadastro de Turma

See merge request tads.cnat/pdsweb/2019.2/g-horarios!54

+++ src/projeto/otime/views.py
@@ -13,15 +13,16 @@ def index(request):
     }
     return render(request, "otime/index.html", contexto)
 
-def reservarHorario(request):
+def reservarHorario(request, id):
     if request.method == "POST":
-        requisicao_post = request.POST
-        sala_de_aula = get_object_or_404(SalaDeAula, pk = requisicao_post["sala_de_aula"])
-        disciplina = get_object_or_404(Disciplina, pk = requisicao_post["disciplina"])
-        professor = get_object_or_404(Professor, pk = requisicao_post["professor"])
-        slot_de_horario = SlotDeHorario(posicao = requisicao_post["posicao"], sala_de_aula = sala_de_aula, disciplina = disciplina, professor = professor)
-        slot_de_horario.save()
+        turma = get_object_or_404(Turma, pk = id)
+        sala_de_aula = get_object_or_404(SalaDeAula, pk = request.POST["sala_de_aula"])
+        disciplina = get_object_or_404(Disciplina, pk = request.POST["disciplina"])
+        professor = get_object_or_404(Professor, pk = request.POST["professor"])
+        for horario in request.POST.getlist("horarios"):
+            slot_de_horario = SlotDeHorario(turma = turma, posicao = horario, sala_de_aula = sala_de_aula, disciplina = disciplina, professor = professor)
 
+            slot_de_horario.save()
     slots_de_horario = SlotDeHorario.objects.all()
     lista_de_salas = SalaDeAula.objects.all()
     lista_de_disciplinas = Disciplina.objects.all()
@@ -38,8 +39,21 @@ def reservarHorario(request):
 
 
 def salas(request):
+    form = FormSala(request.POST or None)
+    if form.is_valid():
+        form.save()
+
     lista_de_salas = SalaDeAula.objects.all()
-    return render(request,'otime/salas.html',{'lista_de_salas':lista_de_salas})
+    return render(request,'otime/salas.html',{'lista_de_salas':lista_de_salas, 'form': form})
+
+def disciplinas(request):
+    form = FormDisciplina(request.POST or None)
+
+    if form.is_valid():
+        form.save()
+
+    lista_de_disciplinas = Disciplina.objects.all()
+    return render(request,'otime/disciplinas.html',{'lista_de_disciplinas':lista_de_disciplinas, 'form': form})
 
 def professores(request):
     form =  FormProfessor(request.POST or None)
@@ -49,27 +63,33 @@ def professores(request):
     lista_de_professores = Professor.objects.all()
     return render(request,'otime/professores.html',{'lista_de_professores':lista_de_professores, 'form':form})
 
-def disciplinas(request):
-    lista_de_disciplinas = Disciplina.objects.all()
-    return render(request,'otime/disciplinas.html',{'lista_de_disciplinas':lista_de_disciplinas})
+##View para exibir as turmas cadastradas e
+##selecionar a que irá ser definida os horarios na view de reserva de horario.
+def seletor_turmas(request):
+    lista_de_turmas = Turma.objects.all()
+    return render(request,'otime/seletor_turmas.html',{'lista_de_turmas':lista_de_turmas})
 
-def criar_disciplina(request):
-    form = FormDisciplina(request.POST or None)
+def turmas(request):
+    form =  FormTurma(request.POST or None)
+    if form.is_valid():
+        form.save()
+
+    lista_de_turmas = Turma.objects.all()
+    return render(request,'otime/turmas.html',{'lista_de_turmas':lista_de_turmas, 'form':form})
+
+def atualizar_turma(request, id):
+    turma = Turma.objects.get(id=id)
+    form =  FormTurma(request.POST or None, instance=turma)
 
     if form.is_valid():
         form.save()
-        return redirect('disciplinas')
+        return redirect('turmas')
 
-    return render(request, 'otime/disciplina-form.html', {'form': form})
+    if request.method == 'POST':
+        turma.delete()
+        return redirect('turmas')
 
-def criar_sala(request):
-    form = FormSala(request.POST or None)
-
-    if form.is_valid():
-        form.save()
-        return redirect('salas')
-
-    return render(request, 'otime/sala-form.html', {'form': form})
+    return render(request, 'otime/modais/editar-turma.html', {'form': form, 'turma': turma})
 
 def atualizar_professor(request, id):
     professor = Professor.objects.get(id=id)
@@ -79,17 +99,11 @@ def atualizar_professor(request, id):
         form.save()
         return redirect('professores')
 
-    return render(request, 'otime/professores.html', {'form': form, 'professor': professor})
+    if request.method == 'POST':
+        professor.delete()
+        return redirect('professores')
 
-def atualizar_disciplina(request, id):
-    disciplina = Disciplina.objects.get(id=id)
-    form =  FormDisciplina(request.POST or None, instance=disciplina)
-
-    if form.is_valid():
-        form.save()
-        return redirect('disciplinas')
-
-    return render(request, 'otime/disciplina-form.html', {'form': form, 'disciplina': disciplina})
+    return render(request, 'otime/modais/editar-prof.html', {'form': form, 'professor': professor})
 
 def atualizar_sala(request, id):
     sala = SalaDeAula.objects.get(id=id)
@@ -99,34 +113,25 @@ def atualizar_sala(request, id):
         form.save()
         return redirect('salas')
 
-    return render(request, 'otime/sala-form.html', {'form': form, 'sala': sala})
-
-def deletar_professor(request, id):
-    professor = Professor.objects.get(id=id)
-
     if request.method == 'POST':
-        professor.delete()
-        return redirect('professores')
+        sala.delete()
+        return redirect('salas')
 
-    return render(request, 'otime/prof-delete-confirm.html', {'professor': professor})
+    return render(request, 'otime/modais/editar-sala.html', {'form': form, 'sala': sala})
 
-def deletar_disciplina(request, id):
+def atualizar_disciplina(request, id):
     disciplina = Disciplina.objects.get(id=id)
+    form =  FormDisciplina(request.POST or None, instance=disciplina)
+
+    if form.is_valid():
+        form.save()
+        return redirect('disciplinas')
 
     if request.method == 'POST':
         disciplina.delete()
         return redirect('disciplinas')
 
-    return render(request, 'otime/disc-delete-confirm.html', {'disciplina': disciplina})
-
-def deletar_sala(request, id):
-    sala = SalaDeAula.objects.get(id=id)
-
-    if request.method == 'POST':
-        sala.delete()
-        return redirect('salas')
-
-    return render(request, 'otime/sala-delete-confirm.html', {'sala': sala})
+    return render(request, 'otime/modais/editar-disc.html', {'form': form, 'disciplina': disciplina})
 
 def modelo(request):
     lista_de_professores = Professor.objects.all()
